% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_summary.R
\name{give_summary}
\alias{give_summary}
\title{Creating summary report from the simulated imputation framework}
\usage{
give_summary(impute_simulated_output)
}
\arguments{
\item{impute_simulated_output}{The saved R object from the \code{\link{impute_simulated}} function (see examples)}
}
\value{
Pdf output with graphs and summary statistics of the performance of the missing data algorithms
}
\description{
\code{\link{give_summary}}
}
\details{

}
\examples{
\dontrun{
give_summary(wrap)
}

}
