% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAR.R
\name{MCAR}
\alias{MCAR}
\title{Missing data spike-in in MCAR pattern}
\usage{
MCAR(X_hat, missfrac_per_var)
}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (Simulated_matrix output from the \code{\link{simulate}} function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the \code{\link{get_data}} function)}
}
\value{
\item{MCAR_matrix}{Matrix with MCAR pred-defined missingess pattern}
\item{Summary}{Summary of MCAR_matrix including number of missing values per variable}
}
\description{
\code{\link{MCAR}} spikes in missingness using missing-completely-at-random (MCAR) pattern
}
\details{
This function uses the generated simulated matrix and generates missing datapoints in a missing-completely-at-random
pattern for each variable, considering the fraction of missingness for each variable, so potential missing data fraction
imbalances between variables in the original data will be retained. The missing data spike-in is completely at random.
}
\examples{
\dontrun{
MCAR(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable)
MCAR(simulated$Simulated_matrix, metadata$Fraction_missingness_per_variable, window = 0.2)
}

}
