% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNAR.R
\name{MNAR}
\alias{MNAR}
\title{Missing data spike-in in MNAR pattern}
\usage{
MNAR(X_hat, MD_pattern, NA_fraction, min_PDM = 10)
}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (Simulated_matrix output from the \code{\link{simulate}} function)}

\item{MD_pattern}{Missing data pattern in the original dataset (MD_Pattern output from the \code{\link{get_data}} function)}

\item{NA_fraction}{Fraction of missingness in the original dataset (Fraction_missingness output from the \code{\link{get_data}} function)}

\item{min_PDM}{All patterns with number of observations less than this number will be removed from the missing data generation. This argument is necessary to be carefully set, as the function will fail or generate erroneous missing data patterns with very complicated missing data patterns. The default is 10, but for large datasets this number needs to be set higher to avoid errors.}
}
\value{
\item{MNAR_matrix}{Matrix with MNAR pred-defined missingess pattern}
\item{Summary}{Summary of MNAR_matrix including number of missing values per variable}
}
\description{
\code{\link{MNAR}} spikes in missingness using missing-not-at-random (MNAR) pattern
}
\details{
This function uses the generated simulated matrix and generates missing datapoints in a missing-not-at-random
pattern for each variable using the \code{\link[mice]{ampute}} function, considering the fraction of missingness in
the original dataset and the original missingness pattern. The characteristic of the MNAR pattern is that
the missingness in a variable is dependent on its own distribution. It is suggested that the user carefully
examines the missing data fractions, excludes variables with high missingess using the \code{\link{clean}} function before proceeding to
missing data generation.
}
\examples{
\dontrun{
MNAR(simulated$Simulated_matrix,
    MD_pattern = metadata$MD_Pattern,
    NA_fraction = metadata$Fraction_missingness,
    min_PDM = 10)
}

}
