% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MAR.R
\name{predict_MAR}
\alias{predict_MAR}
\title{Prediction of MAR missing data pattern based on available data}
\usage{
predict_MAR(X)
}
\arguments{
\item{X}{Number of rows (samples) in the original dataframe (Rows output from the \code{\link{get_data}} function)}
}
\value{
\item{MAR_stats}{xxxxx}
\item{Plot}{xxxxxx}
}
\description{
\code{\link{predict_MAR}} predict missingness status (missing vs. non-missing) based on other variables.
}
\details{
This function xxxx
}
\examples{
\dontrun{
predicted <- predict_MAR(X=df)
}

}
