% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_missing.R
\name{predict_missing}
\alias{predict_missing}
\title{Prediction of missing data based on available data}
\usage{
predict_missing(rownum, colnum, cormat, missfrac_per_var, bootstrap)
}
\arguments{
\item{X_orig}{Dataframe - the original data that contains missing values.}

\item{X_imp}{Dataframe - the imputed data with no missing values.}

\item{scale}{Boolean with default TRUE. Scaling will scale and center all variables to mean=0 and standard deviation=1 in the \strong{original dataframe with missingness}. The user should select TRUE or FALSE here depending on whether the imputed dataframe has scaled or unscaled values.}
}
\value{
\item{Densityplots}{List of density plots of all variables. The density plots show the original values and the imputed values overlaid for each variables in the dataframe}
\item{Boxplot}{List of boxplots of all variables. The boxplots show the original values and the imputed values for each variables in the dataframe. As normally, the boxplots show the median values, the IQR and the range of values}
}
\description{
\code{\link{predict_missing}} predict missingness status (missing vs. non-missing) based on other variables.
}
\details{
This function uses the original dataframe and extracts descriptive meta-data.
}
\examples{
\dontrun{
plots <- post_imp_diag(X_orig = df_miss, X_imp = df_imputed, scale=T)
plots$Densityplots$variable_X
plots$Boxplots$variable_Z
}

}
