% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mean_imp.R
\name{test_mean_imp}
\alias{test_mean_imp}
\title{Testing the mean imputation algorithm}
\usage{
test_mean_imp(X_hat, list)
}
\arguments{
\item{X_hat}{Simulated matrix with no missingess (this matrix will be used to obtain the error between the original and imputed values). (Simulated_matrix output from the \code{\link{simulate}} function)}

\item{list}{List of matrices with various missingness patterns (MCAR, MAR, MNAR and optionally, MAP). (The input is ideally the R object that was generated using the \code{\link{all_patterns}} function)}
}
\value{
\item{Comp_time}{Computation time of imputation using method (default output)}
\item{MCAR_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MCAR missingess pattern (default output)}
\item{MAR_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MAR missingess pattern (default output)}
\item{MNAR_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MNAR missingess pattern (default output)}
\item{MAP_RMSE}{Root-mean-square error (RMSE) between the indexed original values and the imputed values in an MAP missingess pattern (optional output)}
\item{MCAR_MAE}{Mean absolute error (MAE) between the indexed original values and the imputed values in an MCAR missingess pattern (default output)}
\item{MAR_MAE}{Mean absolute error (MAE) between the indexed original values and the imputed values in an MAR missingess pattern (default output)}
\item{MNAR_MAE}{Mean absolute error (MAE) between the indexed original values and the imputed values in an MNAR missingess pattern (default output)}
\item{MAP_MAE}{Mean absolute error (MAE) between the indexed original values and the imputed values in an MAP missingess pattern (optional output)}
\item{MCAR_KS}{Kolmogorov–Smirnov test statistic (KS) between the indexed original values and the imputed values in an MCAR missingess pattern (default output)}
\item{MAR_KS}{Kolmogorov–Smirnov test statistic (KS) between the indexed original values and the imputed values in an MAR missingess pattern (default output)}
\item{MNAR_KS}{Kolmogorov–Smirnov test statistic (KS) between the indexed original values and the imputed values in an MNAR missingess pattern (default output)}
\item{MAP_KS}{Kolmogorov–Smirnov test statistic (KS) between the indexed original values and the imputed values in an MAP missingess pattern (optional output)}
}
\description{
\code{\link{test_mean_imp}} tests the imputation accuracy of the mean imputation algorithm on matrices with various missing data patterns
}
\details{
This function tests the imputation accuracy of the mean imputation algorithm by comparing the
original simulated matrix with no missingess and the imputed matrices generated by the algorithm using the matrices with
MCAR, MAR, MNAR and (optionally) MAP missingness patterns. The function calculates root-mean-square error (RMSE),
mean absolute error (MAE), Kolmogorov–Smirnov D test statistic (KS) between the imputed datapoints and the original
datapoints (that were subsequently set to missing). The function will also calculate the cumulative computation time for
imputing all datasets. The function will automatically detect whether there is a MAP matrix in the list and calculate
RMSE for all matrices provided in the list.
}
\examples{
\dontrun{
test_mean_imp(X_hat = simulated$Simulated_matrix, list = miss_list)
}

}
