% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MAR.R
\name{predict_MAR}
\alias{predict_MAR}
\title{Prediction of MAR missing data pattern based on available data}
\usage{
predict_MAR(X, scale = T)
}
\arguments{
\item{X}{Dataframe - the original data that contains missing values.}

\item{scale}{Boolean with default TRUE. Scaling will scale and center all variables to mean=0 and standard deviation=1}
}
\value{
\item{MAR_stats}{ROC AUC values and 95\% CIs for each variable in the dataset. The ROC AUC statistics are established using all other variables.}
\item{Plot}{Plot visualizing ROC AUC values and 95\% CIs for each variable in the dataset.}
}
\description{
\code{\link{predict_MAR}} predict missingness status (missing vs. non-missing) based on other variables.
}
\details{
This function aims to predict whether a feature approximates a missing-at-random (MAR)
missingness pattern. The function iterates over all variables by converting it to a binary
variable by missing/non-missing status. Following this transformation, all other variables
are used to establish a prediction model (using GLM). ROC AUC statistics are extracted
to establish the collective predictive accuracy of all remaining variables in relation to
the transformed binary feature in question.
}
\examples{
\dontrun{
predicted <- predict_MAR(X=df)
}

}
