% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MAR.R
\name{predict_MAR}
\alias{predict_MAR}
\title{Prediction of missing data based on available data}
\usage{
predict_MAR()
}
\arguments{
\item{rownum}{Number of rows (samples) in the original dataframe (Rows output from the \code{\link{get_data}} function)}
}
\value{
\item{MCAR_AUC}{xxxxx}
\item{MAR_AUC}{xxxxx}
\item{MNAR_AUC}{xxxxx}
\item{Plot}{xxxxxx}
}
\description{
\code{\link{predict_MAR}} predict missingness status (missing vs. non-missing) based on other variables.
}
\details{
This function xxxx
}
\examples{
\dontrun{
xxxx
}

}
