% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_missing.R
\name{predict_missing}
\alias{predict_missing}
\title{Prediction of missing data based on available data}
\usage{
predict_missing(rownum, colnum, cormat, missfrac_per_var, bootstrap)
}
\arguments{
\item{rownum}{Number of rows (samples) in the original dataframe (Rows output from the \code{\link{get_data}} function)}

\item{colnum}{Number of rows (variables) in the original dataframe (Columns output from the \code{\link{get_data}} function)}

\item{cormat}{Correlation matrix of the original dataframe (Corr_matrix output from the \code{\link{get_data}} function)}

\item{missfrac_per_var}{Fraction of missingness per variable (Fraction_missingness_per_variable output from the \code{\link{get_data}} function)}

\item{bootstrap}{Number of bootstraps used to generate ROC AUC statistics.}
}
\value{
\item{MCAR_AUC}{xxxxx}
\item{MAR_AUC}{xxxxx}
\item{MNAR_AUC}{xxxxx}
\item{Plot}{xxxxxx}
}
\description{
\code{\link{predict_missing}} predict missingness status (missing vs. non-missing) based on other variables.
}
\details{
This function xxxx
}
\examples{
\dontrun{
xxxx
}

}
